Split fuzz test into separate file.
