+++ tests/envs/llvm/BUILD
@@ -127,6 +127,17 @@ py_test(
 )
 
 py_test(
+    name = "fork_env_fuzz_test",
+    timeout = "short",
+    srcs = ["fork_env_fuzz_test.py"],
+    deps = [
+        ":fixtures",
+        "//compiler_gym/envs",
+        "//tests:test_main",
+    ],
+)
+
+py_test(
     name = "fresh_environment_observation_reward_test",
     srcs = ["fresh_environment_observation_reward_test.py"],
     shard_count = 8,
