+++ tests/envs/llvm/fork_env_test.py
@@ -3,8 +3,6 @@
 # This source code is licensed under the MIT license found in the
 # LICENSE file in the root directory of this source tree.
 """Tests for LlvmEnv.fork()."""
-from time import time
-
 from compiler_gym.envs import LlvmEnv
 from compiler_gym.util.runfiles_path import runfiles_path
 from tests.test_main import main
@@ -66,40 +64,6 @@ def test_fork_custom_benchmark(env: LlvmEnv):
         assert ir(env) == ir(new_env)
     finally:
         new_env.close()
-
-
-FUZZ_TIME_SECONDS = 5
-A_FEW_RANDOM_ACTIONS = 10
-
-
-def test_fork_state_fuzz_test(env: LlvmEnv):
-    """Run random episodes and check that fork() produces equivalent state."""
-    end_time = time() + FUZZ_TIME_SECONDS
-    while time() < end_time:
-        env.reset()
-
-        # Take a few warmup steps to get an environment in a random state.
-        for _ in range(A_FEW_RANDOM_ACTIONS):
-            _, _, done, _ = env.step(env.action_space.sample())
-            if done:  # Broken episode, restart.
-                break
-        else:
-            # Fork the environment and check that the states are equivalent.
-            new_env = env.fork()
-            try:
-                assert env.state == new_env.state
-                # Check that environment states remain equal if identical
-                # subsequent steps are taken.
-                for _ in range(A_FEW_RANDOM_ACTIONS):
-                    action = env.action_space.sample()
-                    _, _, done_a, _ = env.step(action)
-                    _, _, done_b, _ = new_env.step(action)
-                    assert done_a == done_b
-                    if done_a:  # Broken episode, restart.
-                        break
-                    assert env.state == new_env.state
-            finally:
-                new_env.close()
 
 
 if __name__ == "__main__":
